argument reorder

+++ src/main/java/ee/uustal/ims/persistence/repository/PlayerRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PlayerRepository extends JpaRepository<Player, Integer> {
+public interface PlayerRepository extends JpaRepository<Player, Long> {
 
     @Query("SELECT p FROM Player p WHERE p.username = ?1")
     Optional<Player> findByUserName(String username);
